+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/spells/FillSpell.java
@@ -51,18 +51,9 @@ public class FillSpell extends BrushSpell
 			
 			return SpellResult.CAST;
 		}
-
-		Double ftxValue = parameters.getDouble("ftx", null);
-		Double ftyValue = parameters.getDouble("fty", null);
-		Double ftzValue = parameters.getDouble("ftz", null);
-		if (ftxValue != null && ftzValue != null && ftyValue != null) {
-			Location targetLocation = getLocation();
-			targetLocation = new Location(targetLocation.getWorld(), 
-					ftxValue, 
-					ftyValue, 
-					ftzValue,
-					targetLocation.getYaw(), targetLocation.getPitch());
-			this.targetBlock = targetLocation.getBlock();
+		
+		if (targetLocation2 != null) {
+			this.targetBlock = targetLocation2.getBlock();
 		}
 
 		if (this.targetBlock != null)